Refactor day 7 to use a simpler structure
values are either maps or integers. integer when file. keys are always name of dirs or files

could go EVEN simpler and flatten the structure by representing directory paths as the keys of a map and their values being just the number. as one reads in files, you simply update the structure adding in the new size